Move Image Analysis to ImageAnalyzer component
